Add get, post & delete requests for the pizza

+++ server/app.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 from models import db, Restaurant, RestaurantPizza, Pizza
 from flask_migrate import Migrate
-from flask import Flask, request, make_response, jsonify
+from flask import Flask, request, make_response, jsonify, abort
 from flask_restful import Api, Resource
 import os
 
@@ -32,7 +32,7 @@ def get_restaurants():
 
 @app.route("/restaurants/<int:id>", methods=["GET"])
 def get_restaurant(id):
-    restaurant = Restaurant.query.get(id)
+    restaurant = db.session.query(Restaurant).get(id)
     if not restaurant:
         return jsonify({"error": "Restaurant not found"}), 404
     
@@ -57,11 +57,62 @@ def get_restaurant(id):
 
 @app.route("/restaurants/<int:id>", methods=["DELETE"])
 def delete_restaurant(id):
-    restaurant = Restaurant.query.get_or_404(id)
+    restaurant = db.session.query(Restaurant).get(id)
+    if not restaurant:
+        return jsonify({"error": "Restaurant not found"}), 404
+    
     db.session.delete(restaurant)
     db.session.commit()
     return '', 204
 
+@app.route("/pizzas", methods=["GET"])
+def get_pizzas():
+    pizzas = Pizza.query.all()
+    pizzas_data = [
+        {
+            "id": pizza.id,
+            "ingredients": pizza.ingredients,
+            "name": pizza.name
+        } for pizza in pizzas
+    ]
+    return jsonify(pizzas_data)
+
+@app.route('/restaurant_pizzas', methods=['POST'])
+def add_restaurant_pizza():
+    data = request.get_json()
+    price = data.get('price')
+    pizza_id = data.get('pizza_id')
+    restaurant_id = data.get('restaurant_id')
+
+    if not (price and pizza_id and restaurant_id):
+        return make_response( {"errors": ["validation errors"]}, 400)
+
+    try:
+        new_restaurant_pizza = RestaurantPizza(price=price, pizza_id=pizza_id, restaurant_id=restaurant_id)
+        db.session.add(new_restaurant_pizza)
+        db.session.commit()
+
+        restaurant = Restaurant.query.filter(Restaurant.id==restaurant_id).first()
+        pizza = Pizza.query.filter(Pizza.id==pizza_id).first()
+        response_data = {
+            "id": new_restaurant_pizza.id,
+            "price": new_restaurant_pizza.price,
+            "pizza_id": new_restaurant_pizza.pizza_id,
+            "restaurant_id": new_restaurant_pizza.restaurant_id,
+            "pizza": {
+                "id": pizza.id,
+                "name": pizza.name,
+                "ingredients": pizza.ingredients
+            },
+            "restaurant": {
+                "id": restaurant.id,
+                "name": restaurant.name,
+                "address": restaurant.address
+            }
+        }
+        return make_response(response_data, 201)
+    except ValueError as e:
+        return make_response( {"errors": ["validation errors"]}, 400)
 
 if __name__ == "__main__":
     app.run(port=5555, debug=True)
